*新增聊天功能

+++ Kooco_TexasHoldem/Assets/Resources/QuestSystem/Script/QuestInfo.cs
@@ -8,7 +8,7 @@ public class QuestInfo
     public int GetPoint;            //獲得點數
     public int FinishProgress;      //完成進度
     public int CurrentProgress;     //當前任務進度
-    public int GetCoin;             //完成任務後獲取籌碼
+    public double GetCoin;          //完成任務後獲取籌碼
 }
 
 public enum QuestEnum
